Confusion at one line
Dunno if you accept translation changes, but this one line particularly sounds iffy. You expect that Rika's mom is replying to Rika's dad, but it turns out she is talking to Rika without any way to figure this out from context alone. It would be good if this was in pair with some visual clue, but there is none. This can be fixed in multiple ways. I propose only one of solutions. You can also change it to not hint at all this was said to Rika, f.e.
```
Rika made the curry with me, while her father did nothing!
```

Choice is yours.